Fix Moment and backwards compatibility issues
