+++ docs/interfaces/colorobject.html
@@ -960,7 +960,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">background<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">undefined</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">string</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L1094">index.ts:1094</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L1094">util.ts:1094</a></li>
 					</ul>
 				</aside>
 			</section>
@@ -970,7 +970,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">border<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">undefined</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">string</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L1095">index.ts:1095</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L1095">util.ts:1095</a></li>
 					</ul>
 				</aside>
 			</section>
@@ -980,7 +980,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">highlight<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">undefined</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">object</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L1100">index.ts:1100</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L1100">util.ts:1100</a></li>
 					</ul>
 				</aside>
 			</section>
@@ -990,7 +990,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">hover<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">undefined</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">object</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L1096">index.ts:1096</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L1096">util.ts:1096</a></li>
 					</ul>
 				</aside>
 			</section>
