+++ docs/interfaces/fullcolorobject.html
@@ -960,7 +960,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">background<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L1106">index.ts:1106</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L1106">util.ts:1106</a></li>
 					</ul>
 				</aside>
 			</section>
@@ -970,7 +970,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">border<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L1107">index.ts:1107</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L1107">util.ts:1107</a></li>
 					</ul>
 				</aside>
 			</section>
@@ -980,7 +980,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">highlight<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">object</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L1112">index.ts:1112</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L1112">util.ts:1112</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-type-declaration">
@@ -995,7 +995,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">hover<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">object</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L1108">index.ts:1108</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L1108">util.ts:1108</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-type-declaration">
