+++ docs/interfaces/hsv.html
@@ -963,7 +963,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">h<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L25">index.ts:25</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L25">util.ts:25</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -978,7 +978,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">s<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L29">index.ts:29</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L29">util.ts:29</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -993,7 +993,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">v<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L33">index.ts:33</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L33">util.ts:33</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
