+++ docs/interfaces/rgb.html
@@ -963,7 +963,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">b<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L51">index.ts:51</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L51">util.ts:51</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -978,7 +978,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">g<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L47">index.ts:47</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L47">util.ts:47</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -993,7 +993,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">r<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L43">index.ts:43</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L43">util.ts:43</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
