+++ docs/interfaces/rgba.html
@@ -967,7 +967,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">a<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L73">index.ts:73</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L73">util.ts:73</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -982,7 +982,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">b<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L69">index.ts:69</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L69">util.ts:69</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -997,7 +997,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">g<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L65">index.ts:65</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L65">util.ts:65</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -1012,7 +1012,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">r<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/3f2ce18/src/index.ts#L61">index.ts:61</a></li>
+						<li>Defined in <a href="https://github.com/visjs/vis-util/blob/8da6d10/src/util.ts#L61">util.ts:61</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
